requirements.txt
Pillow==4.1.1
pytest==3.0.7
